+++ src/components/layout/sidebar/index.ts
@@ -1,3 +1,0 @@
-export { SidebarLayout } from "./sidebar-layout";
-export { SidebarNav } from "./sidebar-nav";
-