+++ src/components/layout/topbar/index.ts
@@ -1,3 +1,0 @@
-export { TopbarLayout } from "./topbar-layout";
-export { TopbarNav } from "./topbar-nav";
-